Automation/requirements.txt
google-cloud-firestore
google-cloud-secret-manager
pytz
tenacity
firebase-admin==6.0.1
requests==2.28.1
python-dotenv==0.21.0
